+++ NEWS.md
@@ -1,3 +1,5 @@
+# ruler (development version)
+
 # ruler 0.2.4
 
 * Maintenance release in reaction to a planned update of `tibble`.