+++ NEWS.md
@@ -1,4 +1,4 @@
-# ruler (development version)
+# ruler 0.3.0
 
 * Update minimum `dplyr` version to be 0.8.0.
 * Introduce dependency on `purrr >= 1.0.0` as a reaction to soft deprecation or `rlang::squash()`.